--- Universties_with_dcc.csv
@@ -1,10 +1,0 @@
-Name
-University of Minnesota-Twin Cities
-Syracuse University
-University of Washington
-University of Illinois Chicago
-Miami University of Ohio
-University of Arizona
-University of North Carolina at Asheville
-Duke University
-The University of Texas at Austin
