--- about-us.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang='en'>
-    <head>
-        <meta charset='utf-8'>
-        <title>Colleges For All</title>
-        <link href="https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Lora:ital,wght@0,400..700;1,400..700&family=Share+Tech+Mono&display=swap" rel="stylesheet">
-        <link rel='stylesheet' href='styles.css'>
-        <link rel='stylesheet' href='about-us.css'>
-        
-    </head>
-    <body>
-        <!--NAV BAR-->
-        <nav id="nav-bar" role="navigation" aria-label="Main Navigation">
-            <div id="title">
-              <h1>Colleges for All</h1>
-              <p class="subtitle"><em>the inclusive search tool</em></p>
-            </div>
-            <div id="links">
-              <a href="index.html">Home Page</a>
-              <a href="filter.html">Find a College</a>
-              <a href="about-us.html">About Us</a>
-            </div>
-        </nav>
-        
-        <!--About Us-->
-        <section class="about-container">
-            <h2 class="section-title">Meet Our Team</h2>
-            <div class="team-grid">
-
-                <!--Khai-->
-                <div class="team-member">
-                    <img src="./img/Khai.jpg" alt="A nonbinary viet person wearing a blue plaid shirt and smiling at the beach." class="team-photo">
-                    <h3 class="member-name">Khai</h3>
-                    <p class="member-bio">Khai is a pianist and programmer who likes creating special web experiences with AI.</p>
-                </div>
-
-                <!--Rukhia-->
-                <div class="team-member">
-                    <img src="./img/Rukhia.jpg" alt="A Somali Muslim woman wearing a black hijab and black dress and smiling." class="team-photo">
-                    <h3 class="member-name">Rukhia </h3>
-                    <p class="member-bio">Rukhia is passionate about solving problems to help others through programming, with a love for reading books in her spare time.</p>
-                </div>
-
-                <!--Danielle-->
-                <div class="team-member">
-                    <img src="./img/Dani.jpg" alt="A white woman with brown hair wearing a black shirt smiling at the camera." class="team-photo">
-                    <h3 class="member-name">Danielle </h3>
-                    <p class="member-bio">Danielle is passionate about developing ethical and accessible technology with impact and engaging with community.</p>
-                </div>
-
-                <!--Juliane-->
-                <div class="team-member">
-                    <img src="./img/Juli.jpg" alt="A Caucasian woman with brown hair wearing a blue top and smiling." class="team-photo">
-                    <h3 class="member-name">Juliane </h3>
-                    <p class="member-bio">Juliane builds tech that makes a difference using AI and data and she's a globetrotter at heart.</p>
-                </div>
-            </div>
-        </section>
-    </body>
-
-</html>
