--- filter.css
@@ -1,259 +1,0 @@
-/* General Styles */
-body {
-  font-family: 'Inter', sans-serif;
-  margin: 0;
-  padding: 0;
-}
-
-.hidden {
-  display: none;
- }
-
-
-
-header {
-  background-color: #B2CCC2;
-  color: #000;
-  padding: 20px 60px;
-  width: 100%;
-  font-family: "Open Sans", sans-serif;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-wrap: wrap;
-  box-sizing: border-box;
-  overflow: hidden;
-}
-
-header h1 {
-  font-size: 24px;
-  font-weight: bold;
-  margin-left: -30px;
-  flex-shrink: 0;
-  white-space: nowrap;
-}
-
-/* Main Layout */
-main {
-  display: flex;
-  padding-top: 20px;
-}
-
-/* Filters */
-.filters {
-  width: 33%;
-  background-color: #f8f8f8;
-  padding: 20px;
-  border-radius: 8px;
-  box-sizing: border-box;
-  margin-right: 20px; 
-  margin-top: -20px; 
-}
-
-.filters label {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin: 5px 0;
-  font-size: 16px;
-}
-
-.filters input[type="checkbox"] {
-  margin: 0;
-  width: 16px;
-  height: 16px;
-}
-
-/* Search Container */
-.search-container {
-  width: 67%;
-  padding: 20px;
-  box-sizing: border-box;
-}
-
-.search-area {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 20px;
-}
-
-#search {
-  flex-grow: 1;
-  padding: 12px;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  max-width: 600px;
-}
-
-#search-btn {
-  padding: 12px 15px;
-  background-color: #8aa7a2;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  margin-left: 10px;
-}
-
-#search-btn:hover {
-  background-color: #6e9188;
-}
-
-/* Carousel */
-.college-carousel {
-  display: flex;
-  justify-content: center;
-  gap: 20px;
-  margin-top: 120px;
-}
-.college-carousel.hidden {
-  display: none; 
-}
-
-.college-card {
-  width: 25%;
-  height: 180px;
-  padding: 15px;
-  border: 1px solid #ddd;
-  border-radius: 10px;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
-  text-align: center;
-  background: white;
-  font-size: 16px;
-  font-weight: bold;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  transition: transform 1s ease-in-out;
-}
-
-.college-card.large {
-  width: 35%;
-  height: 200px;
-  transform: scale(1.1);
-  z-index: 2;
-}
-
-.college-card.small {
-  width: 20%;
-  height: 190px;
-  opacity: 0.8;
-}
-
-/* Search Results */
-.college-grid {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-}
-
-.college-card2 {
-  width: 80%;
-  background-color: white;
-  padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-  text-align: left;
-  transition: transform 0.3s;
-  margin-bottom: 20px;
-}
-
-.college-card2:hover {
-  transform: scale(1.02);
-}
-
-.college-card2 h3 {
-  margin: 10px 0;
-  font-size: 20px;
-  font-weight: bold;
-}
-
-.college-card2 p {
-  margin: 5px 0;
-  font-size: 16px;
-  color: #555;
-}
-
-/* Responsive Design */
-@media (max-width: 768px) {
-  main {
-    flex-direction: column;
-  }
-
-  .filters {
-    width: 100%;
-    margin-right: 0;
-    margin-bottom: 20px;
-  }
-
-  .search-container {
-    width: 100%;
-  }
-
-  .college-carousel {
-    flex-direction: column;
-    align-items: center;
-    gap: 15px;
-  }
-
-  .college-card {
-    width: 80%;
-    max-width: 250px;
-    height: 180px;
-  }
-
-  .college-card.large {
-    max-width: 300px;
-    height: 200px;
-  }
-  .college-card.small {
-    width: 80%;
-    max-width: 220px;
-    height: 150px;
-  }
-}
-
-
-#clear-filters-btn {
-  background-color: #8FC3BA; /* Red color to indicate reset */
-  color: black;
-  padding: 12px 20px;
-  font-size: 16px;
-  border: none;
-  border-radius: 8px;
-  cursor: pointer;
-  display: block;
-  margin-top: 15px;
-  width: 60%;
-}
-
-#clear-filters-btn:hover {
-  background-color: black;
-  color: white 
-}
-
-
-#selectState {
-  font-size: 18px; 
-  padding: 12px; 
-  border-radius: 8px; 
-  border: 2px solid #555; 
-  width: 100%; 
-  max-width: 350px; 
-  background-color: white;
-  cursor: pointer;
-  outline: none;
-}
-
-
-#selectState:focus {
-  border-color: #648882;
-}
-
-#selectState option {
-  font-size: 16px;
-  padding: 10px;
-}
