--- filter.html
@@ -1,151 +1,0 @@
-<!--
-  College Inclusiveness Search Team
-  February 2025
--->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Colleges for All</title>
-    <link href="https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Lora:ital,wght@0,400..700;1,400..700&family=Share+Tech+Mono&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="styles.css">
-    <link rel="stylesheet" href="filter.css">
-</head>
-<body>
-  <nav id="nav-bar" role="navigation" aria-label="Main Navigation">
-    <div id="title">
-      <h1>Colleges for All</h1>
-      <p class="subtitle"><em>the inclusive search tool</em></p>
-    </div>
-    <div id="links">
-      <a href="index.html">Home Page</a>
-      <a href="filter.html">Find a College</a>
-      <a href="about-us.html">About Us</a>
-    </div>
-  </nav>
-      <!----
-          INSERT SCREEN-READER HERE
-      -->
-    <main>
-        <aside class="filters">
-            <h3>What's important for your college to have?</h3>
-            <p>You can filter by one or multiple categories (such as state, accommodations, and identity)</p>
-            <br/>
-            <legend>Filter by State</legend>
-            <br/>
-            <select id="selectState">
-              <option value= "None">None</option>
-              <option value="AL">Alabama</option>
-              <option value="AK">Alaska</option>
-              <option value="AZ">Arizona</option>
-              <option value="AR">Arkansas</option>
-              <option value="CA">California</option>
-              <option value="CO">Colorado</option>
-              <option value="CT">Connecticut</option>
-              <option value="DE">Delaware</option>
-              <option value="DC">District Of Columbia</option>
-              <option value="FL">Florida</option>
-              <option value="GA">Georgia</option>
-              <option value="HI">Hawaii</option>
-              <option value="ID">Idaho</option>
-              <option value="IL">Illinois</option>
-              <option value="IN">Indiana</option>
-              <option value="IA">Iowa</option>
-              <option value="KS">Kansas</option>
-              <option value="KY">Kentucky</option>
-              <option value="LA" >Louisiana</option>
-              <option value="ME">Maine</option>
-              <option value="MD">Maryland</option>
-              <option value="MA">Massachusetts</option>
-              <option value="MI">Michigan</option>
-              <option value="MN">Minnesota</option>
-              <option value="MS">Mississippi</option>
-              <option value="MO">Missouri</option>
-              <option value="MT">Montana</option>
-              <option value="NE">Nebraska</option>
-              <option value="NV">Nevada</option>
-              <option value="NH">New Hampshire</option>
-              <option value="NJ">New Jersey</option>
-              <option value="NM">New Mexico</option>
-              <option value="NY">New York</option>
-              <option value="NC">North Carolina</option>
-              <option value="ND">North Dakota</option>
-              <option value="OH">Ohio</option>
-              <option value="OK">Oklahoma</option>
-              <option value="OR">Oregon</option>
-              <option value="PA">Pennsylvania</option>
-              <option value="RI">Rhode Island</option>
-              <option value="SC">South Carolina</option>
-              <option value="SD">South Dakota</option>
-              <option value="TN">Tennessee</option>
-              <option value="TX">Texas</option>
-              <option value="UT">Utah</option>
-              <option value="VT">Vermont</option>
-              <option value="VA">Virginia</option>
-              <option value="WA">Washington</option>
-              <option value="WV">West Virginia</option>
-              <option value="WI">Wisconsin</option>
-              <option value="WY">Wyoming</option>
-              </select>
-              <br/><br/><br/>
-            <button id="clear-filters-btn">Clear All Filters</button>
-            <br/>
-            <br/>
-
-            <fieldset>
-            <legend>Filter by Accommodations</legend>
-            <label><input type="checkbox" class="filter-option" value="Accessible Campus"> Accessible Campus</label>
-            <label><input type="checkbox" class="filter-option" value="Accessible Materials"> Accessible Materials</label>
-            <label><input type="checkbox" class="filter-option" value="Notetaking"> Notetaking</label>
-            <label><input type="checkbox" class="filter-option" value="Sign Language Interpreter"> Sign Language Interpreter</label>
-            <label><input type="checkbox" class="filter-option" value="Accessible Parking"> Accessible Parking</label>
-            <label><input type="checkbox" class="filter-option" value="Calm Exam Spaces"> Calm Exam Spaces</label>
-            <label><input type="checkbox" class="filter-option" value="Reliable Accommodations"> Reliable Accommodations</label>
-          </fieldset>
-          <br />
-          <fieldset>
-            <legend>Filter by Identity</legend>
-            <label><input type="checkbox" class="filter-option" value="dcc"> Disability Cultural Center</label>
-            <label><input type="checkbox" class="filter-option" value="Queer-friendly"> Queer-friendly</label>
-            <label><input type="checkbox" class="filter-option" value="Neurodivergent-friendly"> Neurodivergent-friendly</label>
-            <label><input type="checkbox" class="filter-option" value="Religion-inclusive"> Religion-inclusive</label>
-            <label><input type="checkbox" class="filter-option" value="BIPOC-inclusive"> BIPOC-inclusive</label>
-            <label><input type="checkbox" class="filter-option" value="Disability-friendly"> Disability-friendly</label>
-          </fieldset>
-        </aside>
-
-        <div class="search-container">
-          <p id="search-status" aria-live="polite"></p>
-          <div id="search-bar">
-            <input type="text" aria-label="Search bar" id="search"list="suggestions" placeholder="type in a college here...">
-            <datalist id="suggestions"></datalist>
-            <button class="btn">Search</button>
-            <br/><br/><br/><br/><br/>
-          </div>
-
-
-          <section id="college-carousel" class="college-carousel" aria-live="polite">
-
-            <!--Left card, not focused--> 
-            <div class="college-card small" id="left-card" tabindex="-1" role="article" 
-              aria-label="College preview left" aria-hidden="true">
-            </div>
-            <!--Center card, focused-->
-            <div class="college-card large" id="center-card" tabindex="0" role="article" aria-label="Main college preview"></div>
-            
-            <!--Right card, not focused--> 
-            <div class="college-card small" id="right-card" tabindex="-1" role="article" 
-              aria-label="College preview right" aria-hidden="true">
-            </div>
-          </section>
-
-          <div id="results" class="hidden college-grid">
-      </div>
-
-
-    </main>
-
-    <script src="filter.js"></script>
-</body>
-</html>
