--- index.html
@@ -1,91 +1,0 @@
-<!--
-  College Inclusiveness Search Team
-  February 2025
--->
-
-<!DOCTYPE html>
-<html lang='en'>
-  <head>
-    <meta charset='utf-8'>
-    <title>Colleges For All</title>
-    <link href="https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Lora:ital,wght@0,400..700;1,400..700&family=Share+Tech+Mono&display=swap" rel="stylesheet">
-    <link rel='stylesheet' href='styles.css'>
-    <script src='index.js'></script>
-  </head>
-  <body>
-    <!----
-        INSERT SCREEN-READER HERE
-    -->
-    <nav id="nav-bar" role="navigation" aria-label="Main Navigation">
-      <div id="title">
-        <h1>Colleges for All</h1>
-        <p class="subtitle"><em>the inclusive search tool</em></p>
-      </div>
-      <div id="links">
-        <a href="index.html">Home Page</a>
-        <a href="filter.html">Find a College</a>
-        <a href="about-us.html">About Us</a>
-      </div>
-    </nav>
-
-    <div id="search-bar">
-      <input type="text" aria-label="Search bar" list="suggestions" tabindex="0" placeholder="type in a college here..." for="collegeSearch">
-      <datalist id="suggestions"></datalist>
-      <button class="btn">Search</button>
-    </div>
-    <h2 >find the perfect college for your unique identity...</h2>
-    <div id="cards">
-      <div id="disability-card" class="card" role="region">
-        <h3>based on accommodations</h3>
-        <ul aria-label="Accommodations available">
-          <li>Accessible Campus</li>
-          <li>Accessible Materials</li>
-          <li>Reliable Accommodations</li>
-          <li>Sign Language Interpreters</li>
-        </ul>
-      </div>
-      <div id="inclusivity-card" class="card" role="region">
-        <h3>based on inclusivity</h3>
-        <ul aria-label="Inclusivity features">
-          <li>Queer-friendly</li>
-          <li>BIPOC-inclusive</li>
-          <li>Disability-friendly</li>
-          <li>Religion-inclusive</li>
-        </ul>
-      </div>
-    </div>
-    <h2 >what is it like being intersectional in college?</h2>
-    <div id="experiences">
-      <div  class="experience" role="region">
-        <div class="person">
-          <h3>Luisa</h3>
-          <img src="img/luisa.jpg" alt="Clip art of a Femme person in a wheelchair joyfully holding a Pride flag.">
-          <p><em>queer & disabled woman</em></p>
-          <p>Sophomore</p>
-          <p>University of Washington</p>
-        </div>
-        <p>At UW Seattle, I've experienced plenty of ups and downs, no literally the wheelchair accessibility is pretty bad for a super hilly and stair-filled campus. At least the students here are pretty understanding and there are cool RSOs here for queer disabled women like me!</p>
-      </div>
-      <div class="experience" role="region">
-        <div class="person">
-          <h3>Bobby</h3>
-          <img src="img/bobby.jpg" alt="A smiling black man in a wheelchair.">
-          <p ><em>black & disabled man</em></p>
-          <p>Junior</p>
-          <p>Harvard University</p>
-        </div>
-        <p>At Harvard, students don't seem to care much about your disability and instead focus on your intellectual ability -- which is nice but disability shapes my day-to-day too, not everyone will understand the inconveniences of 2nd floor classes and materials that don't work with screenreaders.</p>
-      </div>
-      <div class="experience" role="region">
-        <div class="person">
-          <h3>Honda</h3>
-          <img src="img/honda.jpg" alt="A smiling Asian man wearing glasses and a backpack.">
-          <p><em>asian man with ADHD</em>
-          <br>Freshman
-          <br>Yale University</p>
-        </div>
-        <p>I have more chaotic energy than what the model minority stereotype suggests, but hey at least my friends at Yale are chill w/ that! That being said I really appeciate everyone's understanding of my ADHD needs, like flexible timing and distraction-free environments.</p>
-      </div>
-    </div>
-  </body>
-</html>
