+++ public/college.html
@@ -1,0 +1,126 @@
+<!DOCTYPE html>
+<html lang='en'>
+  <head>
+    <meta charset='utf-8'>
+    <title>Colleges For All</title>
+    <link href="https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Lora:ital,wght@0,400..700;1,400..700&family=Share+Tech+Mono&display=swap" rel="stylesheet">
+    <link rel='stylesheet' href='styles.css'>
+    <link rel='stylesheet' href='college.css'>
+    <script src='college.js'></script>
+  </head>
+  <body>
+    <nav id="nav-bar" role="navigation" aria-label="Main Navigation">
+      <div id="title">
+        <h1>Colleges for All</h1>
+        <p class="subtitle"><em>the inclusive search tool</em></p>
+      </div>
+      <div id="links">
+        <a href="index.html">Home Page</a>
+        <a href="filter.html">Find a College</a>
+        <a href="about-us.html">About Us</a>
+      </div>
+    </nav>
+    <!----
+        INSERT SCREEN-READER HERE
+    -->
+    <div id="college-section">
+      <div id="name-of-college">
+        <h2 id="college-name">
+          <a href="#" id="drs-link" target="_blank">University of Atlantis</a>
+        </h2>
+      </div>
+      <div id="rate-me-button">
+        <button id="rate-btn" type="button" class="button">Rate this college</button>
+      </div>
+    </div>
+    <div id="ratings">
+      <div id="rating-left" aria-label="Overall ratings:">
+        <div id="overall-inclusivity">
+          <p>Overall Inclusivity</p>
+          <span class="overall-display"></span>
+        </div>
+
+        <div id="overall-accessibility">
+          <p>Overall Accessibility</p>
+          <span class="overall-display"></span>
+        </div>
+
+      </div>
+      <div id="rating-middle" aria-label="Average ratings:">
+        <div id="lgbtq-rating">
+          <p>LGBTQ+ Friendliness</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="accommodation-rating">
+          <p >Accommodation Friendliness</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="difficulty-rating">
+          <p >Accommodation Difficulty</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="reliability-rating">
+          <p>Accommodation Reliability</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="outdoor-accessibility">
+          <p>Outdoor Campus Accessibility</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="indoor-accessibility">
+          <p>Indoor Campus Accessibility</p>
+          <span class="rating-display"></span>
+        </div>
+      </div>
+      <div id="rating-right">
+        <div id="peer-support">
+          <p >Peer Support</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="liberal">
+          <p>Liberal Climate</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="cultural-diversity">
+          <p>Cultural & Racial Diversity</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="religious-tolerance">
+          <p>Religious Tolerance</p>
+          <span class="rating-display"></span>
+        </div>
+        <div id="club-inclusiveness">
+          <p>Club Inclusiveness</p>
+          <span class="rating-display"></span>
+        </div>
+      </div>
+    </div>
+    <div id="reviews">
+      <div id="reviews-left">
+        <h2 >What students are saying:</h2>
+        <br>
+        <p><strong>Accommodations and Identity at a glance:</strong></p>
+        <p><strong></strong></p>
+        <ul id="accoms">
+          <li>Sign Language Interpreter</li>
+          <li>Reduced Distraction Environment</li>
+          <li>Accessible Parking</li>
+        </ul>
+        <br>
+        <div id="percent">
+          <p><span id="lgbtq_id"></span>% of students here identify as part of the LGBTQ+</p>
+          <p><span id="exclusionary"></span>% of students say this college is gender inclusive and inclusive of all sexual identities/orientations</p>
+          <p><span id="friendly"></span>% of students say this college is queer-friendly</p>
+          <p><span id="mobility"></span>% of students with mobility issues say this college has inacccessible buildings</p>
+        </div>
+        <!-- <p><strong>Identity</strong></p>
+        <ul>
+          <li>Trans-friendly</li>
+        </ul> -->
+      </div>
+      <div id="reviews-right">
+        <h3><span id="n-ratings">0</span> Ratings</h3>
+      </div>
+    </div>
+  </body>
+</html>
