+++ public/styles.css
@@ -1,0 +1,187 @@
+/*
+ * College Inclusiveness Search Team
+ * February 2025
+ */
+
+body {
+  font-family: "Inter", serif;
+  margin: 0;
+  padding: 0;
+}
+
+#title {
+  margin-top: 10px;
+  margin-bottom: 10px;
+  margin-left: 70px;
+}
+
+h1, .subtitle {
+  margin: 5px;
+}
+
+h1 {
+  font-weight: 400;
+  font-size: 3em;
+}
+
+.subtitle {
+  font-weight: 200;
+  font-size: 2em;
+}
+
+#nav-bar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: #8FC3BA;
+}
+
+#links {
+  display: flex;
+  justify-content: center;
+  margin-right: 100px;
+}
+
+#links a {
+  margin-left: 60px;
+  text-decoration: none;
+  color: black;
+  font-size: 1.5em;
+  transition: 0.2s;
+  padding: 10px;
+  border-radius: 10px;
+}
+
+#links a:visited {
+  color: black;
+}
+
+#links a:hover {
+  background-color: black;
+  color: white;
+}
+
+#search-bar {
+  margin-top: 40px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+#search-bar input {
+  font-family: inherit;
+  border-radius: 10px;
+  padding: 10px;
+  font-size: larger;
+  width: 700px;
+  margin-right: 10px;
+  height: fit-content;
+}
+
+#search-bar .btn {
+  color: black;
+  transition: 0.2s;
+  font-size: 14px;
+  border-radius: 10px;
+  min-width: 100px;
+  height: 45px;
+  padding: 10px;
+  border: 1px solid black;
+  background-color: white;
+}
+
+#search-bar .btn:hover {
+  background-color: black;
+  color: white;
+}
+
+h2 {
+  font-weight: 400;
+  font-size: 3em;
+  text-align: center;
+}
+
+#cards {
+  display: flex;
+  justify-content: center;
+}
+
+.card {
+  border: 1px solid rgb(113, 113, 113);
+  margin: 20px;
+  width: 500px;
+}
+
+.card h3 {
+  font-size: 1.5em;
+  font-weight: 400;
+  padding: 10px;
+  margin: 0;
+  text-align: center;
+}
+
+#inclusivity-card h3 {
+  background-color: rgb(233, 255, 168);
+}
+
+#disability-card h3 {
+  background-color: rgb(255, 198, 255);
+}
+
+.card li {
+  font-size: larger;
+  margin-bottom: 10px;
+}
+
+#experiences {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+.experience {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 50px;
+}
+
+
+.experience .person {
+  background-color: #8FC3BA;
+  padding: 20px;
+  padding-top: 0;
+  width: fit-content;
+  border: gray 1px solid;
+  border-right: none;
+  border-top-left-radius: 10px;
+  border-bottom-left-radius: 10px;
+}
+
+.experience .person * {
+  margin-bottom: 5px;
+  text-align: center;
+}
+
+.experience .person h3 {
+  font-weight: 500;
+  font-size: 1.5em;
+  text-align: center;
+  margin-bottom: 10px;
+}
+
+.experience .person img {
+  border-radius: 20px;
+  width: 200px;
+}
+
+.experience > p {
+  margin: 0;
+  padding: 20px;
+  width: 60%;
+  background-color: azure;
+  border: gray 1px solid;
+  border-left: none;
+  border-top-right-radius: 10px;
+  border-bottom-right-radius: 10px;
+  font-size: larger;
+}
