+++ public/survey.css
@@ -1,0 +1,194 @@
+/*
+ * College Inclusiveness Search Team
+ * February 2025
+ */
+/* Style the form */
+#survey {
+  
+  margin: 100px auto;
+  padding: 10px;
+  width: 80%;
+  min-width: 300px;
+
+}
+
+body {
+  background-color: #8fc3ba88;
+  font-family: "Open Sans", serif;
+  margin-left: 50px;
+}
+
+h1 {
+  font-size: 3ch;
+}
+h2{
+  font-size:2ch;
+}
+
+
+
+/* Style the input fields */
+input.text-entry,textarea {
+  padding: 10px;
+  width: 90%;
+  font-size: 17px;
+  font-family: Raleway;
+  border: 1px solid #aaaaaa;
+}
+input{
+  margin:0.4rem;
+}
+
+/* Mark input boxes that gets an error on validation: */
+input.invalid {
+  background-color: #ffdddd;
+}
+
+fieldset{
+  background-color: #f9f9f9; /* Light background */
+
+  display: flex;
+  flex-direction: column;
+
+  padding: 6.5rem 15px 15px; /* Add padding */
+  padding-bottom: 1rem;
+  margin-top: 30px;
+  margin-bottom: 20px; /* Space between questions */
+  border-radius: 10px; /* Rounded corners */
+  border: 1px solid #ccc; /* Light border */
+  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1); /* Soft shadow */
+  position: relative; /* Ensure proper positioning */
+  /*padding-top: 2.5rem;  Creates space for the legend */
+}
+
+legend {
+  
+  font-size: 1.1rem;
+  padding: 20px 10px; /* Add padding for better spacing */
+  background: #f9f9f9; /* Match fieldset background */
+  border-radius: 5px;
+  position: absolute; /* Fixes floating effect */
+  top: 0; /* Aligns to the top */
+  left: 15px; /* Aligns with fieldset padding */
+  margin-bottom: 4px;
+}
+
+label{
+  font-size:18px;
+}
+
+/* Hide all steps by default: */
+.tab {
+  display: none;
+  
+}
+
+/* Make circles that indicate the steps of the form: */
+.step {
+  height: 15px;
+  width: 15px;
+  margin: 0 2px;
+  background-color: #bbbbbb;
+  border: none;
+  border-radius: 50%;
+  display: inline-block;
+  opacity: 0.5;
+}
+
+/* Mark the active step: */
+.step.active {
+  opacity: 1;
+}
+
+/* Mark the steps that are finished and valid: */
+.step.finish {
+  background-color: #04AA6D;
+}
+
+.btn {
+  display:inline-block;
+  width: 100px;
+  height: 45px;
+  border-radius: 12px;
+  font-weight: bold;
+  font-size: 15px;
+  text-align: center;
+  cursor: pointer;
+  background-color: #8fc3ba;
+}
+
+.btn:hover {
+  background-color: black;
+  color: white;
+}
+
+/*Container around each question: */
+.question, .section {
+  background-color: #f9f9f9; /* Light background */
+
+  display: flex;
+  flex-direction: column;
+
+  padding: 15px; /* Add padding */
+  margin-bottom: 20px; /* Space between questions */
+  border-radius: 10px; /* Rounded corners */
+  border: 1px solid #ccc; /* Light border */
+  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1); /* Soft shadow */
+  
+}
+
+.multi-select {
+  padding-top: 2rem;
+  display: flex;
+  flex-direction: column; /* Stacks checkboxes vertically */
+  gap: 10px; /* Adds spacing between checkboxes */
+}
+
+.multi-select label {
+  display: flex;
+  align-items: center;
+  gap: 8px; /* Space between checkbox and text */
+  font-size: 16px;
+}
+
+
+.star-rating {
+  border: none;
+  margin-bottom: 20px;
+  
+}
+
+.star-rating fieldset{
+  display: flex;
+  /* flex-direction: row-reverse; Right to left for better UX */
+  justify-content: start;
+  gap: 5px;
+}
+
+.star-rating input {
+  opacity: 0;
+  position: absolute;
+}
+
+.star-rating label {
+  font-size: 30px;
+  cursor: pointer;
+  color: #ccc; /* Default unselected color */
+  transition: color 0.2s;
+}
+
+.star-rating label:hover,
+.star-rating label:hover ~ label {
+  color:  rgb(219, 164, 0); /* Hover effect */
+}
+
+.star-rating input:checked ~ label {
+  color:  rgb(219, 164, 0); /* Selected stars stay highlighted */
+}
+
+/* Add visual feedback for focus (when tabbing) */
+.star-rating input:focus + label,
+.star-rating label:focus {
+  color: rgb(219, 164, 0); /* Highlight focused star */
+}
+
